+++ apps/api/src/data/http/repositories/data-sources-repository.ts
@@ -1,12 +1,12 @@
-import { DataSource, DataSourcesRepository } from "@entities/projects";
+import { DataSet, DataSetsRepository } from "@entities/projects";
 import { Injectable } from "@nestjs/common";
 import { createJiraClient } from "../client/jira-client";
-import { findDataSources } from "@agileplanning-io/flow-data";
+import { findDataSets } from "@agileplanning-io/flow-data";
 
 @Injectable()
-export class HttpJiraDataSourcesRepository extends DataSourcesRepository {
-  async getDataSources({ domain, query }): Promise<DataSource[]> {
+export class HttpJiraDataSetsRepository extends DataSetsRepository {
+  async getDataSets({ domain, query }): Promise<DataSet[]> {
     const client = createJiraClient(domain);
-    return findDataSources(client, query);
+    return findDataSets(client, query);
   }
 }