+++ packages/data/index.ts
@@ -1,1 +1,1 @@
-export * from "./src/data-sources";
+export * from "./src/data-sets";