+++ packages/data/src/data-sets.ts
@@ -2,32 +2,29 @@ import { Version3Client } from "jira.js";
 import { getAllPages } from "./page-utils";
 import { flat } from "remeda";
 
-export type DataSource = {
+export type DataSet = {
   name: string;
   type: "filter" | "project";
   jql: string;
 };
 
-export const findDataSources = async (
-  client: Version3Client,
-  query: string,
-) => {
+export const findDataSets = async (client: Version3Client, query: string) => {
   const normalisedQuery = query?.toLowerCase();
 
   const projects = await findProjects(client, query);
   const filters = await findFilters(client, query);
 
-  const dataSources = [...projects, ...filters];
+  const dataSets = [...projects, ...filters];
 
-  return dataSources.filter(
-    (dataSource) => dataSource.name?.toLowerCase().includes(normalisedQuery),
+  return dataSets.filter(
+    (dataSet) => dataSet.name?.toLowerCase().includes(normalisedQuery),
   );
 };
 
 const findFilters = async (
   client: Version3Client,
   query: string,
-): Promise<DataSource[]> => {
+): Promise<DataSet[]> => {
   const filterPages = await getAllPages(async (startAt) =>
     client.filters.getFiltersPaginated({
       filterName: query,
@@ -51,7 +48,7 @@ const findFilters = async (
 const findProjects = async (
   client: Version3Client,
   query: string,
-): Promise<DataSource[]> => {
+): Promise<DataSet[]> => {
   const projectPages = await getAllPages((startAt) =>
     client.projects.searchProjects({
       query,